Move functions into utils pkg
